CWTInputDataPoint data type added. Method for getting list of retention time and intensity added.

+++ msdk-featdet/msdk-featdet-ADAP-3D/src/main/java/io/github/msdk/featdet/ADAP3D/datamodel/SparseMatrixTriplet.java
@@ -17,7 +17,7 @@ package io.github.msdk.featdet.ADAP3D.datamodel;
  * </p>
  */
 public class SparseMatrixTriplet {
-	public double mz;
+	public int mz;
 	public int scanNumber;
 	public float intensity;
 	public float rt;
